Add function to normaöoze bridge id and use it where applicable Some smaller clean up
